+++ src/tpl/arborx/src/details/ArborX_DetailsBatchedQueries.hpp
@@ -163,12 +163,12 @@ public:
   template <typename ExecutionSpace>
   static std::tuple<Kokkos::View<int *, DeviceType>,
                     Kokkos::View<int *, DeviceType>,
-                    Kokkos::View<float *, DeviceType>>
+                    Kokkos::View<double *, DeviceType>>
   reversePermutation(ExecutionSpace const &space,
                      Kokkos::View<unsigned int const *, DeviceType> permute,
                      Kokkos::View<int const *, DeviceType> offset,
                      Kokkos::View<int const *, DeviceType> indices,
-                     Kokkos::View<float const *, DeviceType> distances)
+                     Kokkos::View<double const *, DeviceType> distances)
   {
     auto const tmp_offset = permuteOffset(permute, offset);
